src/python/rerun_logger/pyproject.toml
[project]
name = "rerun-logger"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "dora-rs==0.4.1",
    "pyarrow>=23.0.1",
    "rerun-sdk>=0.21",
]
